man: make it possible to pipe manpage through a preprocessor
the preprocessor name is manpp and expected to live in $PATH.

using an external preprocessor, we can sanitize some of the
220 groff, mandoc, etc macros currently not provided, without
having to hack into the guts of this problem.
the preprocessor can just apply a couple of sed or awk rules
which strip out the junk or translate it into proper nroff.