man: commit some obsolete code
before i added preprocessor support, i added this code that tries
to distinguish between "unknown" macros that end up as a comment
and such that are just printed as regular text.
this turned out not to be right approach, but it gives hints
as to how such a thing could be implemented.
thus i'm committing this for documentation reasons, just to revert
it in the next one.