add join(1) implementation from FreeBSD
i recently came upon a user of join(1), gnulib-tool.
join is a standard POSIX utility, but the 9base implementation we
used in sabotage does not support the POSIX options at all.
then i switched to the version of sbase, but that one is too buggy.
using it with gnulib-tool caused missing output, and subsequently
a failed build due to missing files gnulib-tool was supposed to copy.

first i played around and created an own join prototype in python,
using a unique design with just 2 buffered lines per file, in order
to avoid memory allocation, so the tool is 100% robust.
however it turned out that there are cases where it is unavoidable to
seek back arbitrary numbers of lines, for example given the files
1 a\n1 b\n and 1 1\n1 2\n, which outputs
1 a 1
1 a 2
1 b 1
1 b 2
i.e.: iterate over all lines in file2 that match on the first occurence
of the join field "1" in file1, and then another time for the second.
since stdin can be used as input for one of the 2 files, seeking cannot
be used, so the tool needs to write either temporary files first,
which can fail as well, or use dynamic memory allocation.
so i came to the conclusion that if i cannot improve upon the existing
designs, i can just as well simply adopt an existing one.

i looked at the NetBSD implementation first, but that one is stuck in
the past and causes several test failures with the GNU coreutils test,
and does not support the POSIX -o option fully (0 as the index for the
join field is not supported).

in order to use the testsuite, one needs 3 files from coreutils:
tests/misc/join.pl, the actual test;
tests/Coreutils.pm, some library functions; and
src/getlimits
the latter you can compile in the coreutils tree via: make src/getlimits
then you need to patch Coreutils.pm to run getlimits from whereever you
have it, or add it to PATH. after that modify join.pl to start the join
implementation to test. run it with: perl -I. -MCoreutils join.pl

the first couple of tests are reasonable (up to 10) and test POSIX
behaviour (apart from 3c, which uses zero bytes).
FreeBSD's join is the only one passing them all.
however there's still one test that fails:

['v2-order', '-v2 -2 2', ["", "2 1\n"], "1 2\n", 0],
that means given 2 files, one empty ("") and the other one containing
"2 1\n", using the options -v2 -2 2 the tool should ouput 1 2, not 2 1.

this join implementation was taken from
https://raw.githubusercontent.com/freebsd/freebsd/master/usr.bin/join/join.c
as of commit 7e6cabd06e6caa6a02eeb86308dc0cb3f:
"Renumber copyright clause 4".

the only things changed are one instance of usage of errc(), removal of
some nonportable junk headers like sys/cdefs.h, and added inclusion of
sys/types.h.